fix: rename matrix_mul.cpp to matrix.hpp
