+++ hw5/Yojahuang/Makefile
@@ -1,11 +1,12 @@
 CXX = g++
-CFLAGS = -std=c++11 -O3 -Wall -shared -fPIC `python3-config --includes` 
+CFLAGS = -std=c++14 -O3 -Wall -shared -fPIC `python3-config --includes` 
 LDFLAGS= -lblas
+CIMPORTPATH = "includes"
 
 all: _matrix.so
 
-_matrix.so: matrix_mul.cpp ../mod.cpp
-	$(CXX) $^ $(CFLAGS) -I/usr/include/mk -o $@ ${LDFLAGS}
+_matrix.so: srcs/Matrix.cpp srcs/Matrix_Mul.cpp includes/matrix.hpp ../mod.cpp
+	$(CXX) $^ $(CFLAGS) -I/usr/include/mk -I${CIMPORTPATH} -o $@ ${LDFLAGS}
 
 clean:
 	rm -f *.so
